update for new design of api telemetry
