+++ change/@microsoft-teams-js-24933635-de42-4b04-b53f-935c1799929a.json
@@ -1,6 +1,6 @@
 {
   "type": "patch",
-  "comment": "Added additional apiVersionTag argument on App, Dialog, GeoLocation, Location, Navigation, Pages, Tasks capabilities for telemetry",
+  "comment": "Added additional telemetry to `App`, `Dialog`, `GeoLocation`, `Location`, `Navigation`, `Pages`, and `Tasks` capabilities",
   "packageName": "@microsoft/teams-js",
   "email": "kangxuanye@microsoft.com",
   "dependentChangeType": "patch"
