+++ packages/teams-js/src/public/location.ts
@@ -1,6 +1,7 @@
-import { getApiVersionTag, sendMessageToParentWithVersion } from '../internal/communication';
-import { ApiName, ApiVersionNumber, locationAPIsRequiredVersion } from '../internal/constants';
+import { sendMessageToParentWithVersion } from '../internal/communication';
+import { locationAPIsRequiredVersion } from '../internal/constants';
 import { ensureInitialized, isCurrentSDKVersionAtLeast } from '../internal/internalAPIs';
+import { ApiName, ApiVersionNumber, getApiVersionTag } from '../internal/telemetry';
 import { errorNotSupportedOnPlatform, FrameContexts } from './constants';
 import { ErrorCode, SdkError } from './interfaces';
 import { runtime } from './runtime';
