+++ packages/teams-js/test/public/navigation.spec.ts
@@ -1,8 +1,8 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
-import * as pagesUtil from '../../src/internal/pagesUtil';
 import * as utilFunc from '../../src/internal/utils';
 import { app, FrameContexts } from '../../src/public';
 import { navigateBack, navigateCrossDomain, navigateToTab, returnFocus } from '../../src/public/navigation';
+import * as pagesUtil from '../../src/public/pages';
 import { Utils } from '../utils';
 
 /* eslint-disable */
